Updated Documentation to Explain Args Usage in NS2
The documentation does not explain how NS2 scopes the game arguments, I wasted 30 minutes trying to figure out why I couldn't get args to work as this file stated. It's explained in other files, but this one looks like it missed the updating.

Simple fix, I suspect this is just a minor oversight, code now includes an example of how to use it with ns2.